move SQS, SNS & lambda to product service
