+++ import-service/package.json
@@ -12,11 +12,9 @@
   "license": "ISC",
   "dependencies": {
     "aws-sdk": "^2.991.0",
-    "axios": "^0.21.4",
     "csv-parser": "^3.0.0",
     "lodash": "^4.17.21",
-    "winston": "^3.3.3",
-    "yup": "^0.32.9"
+    "winston": "^3.3.3"
   },
   "devDependencies": {
     "@types/jest": "^27.0.1",
