+++ product-service/addProduct/addProduct.js
@@ -7,26 +7,25 @@ import { PostgresClient } from "../PostgresClient";
 
 export const addProduct = async event => {
   log.info(`addProduct function is called with args: ${util.inspect(event)}`);
-  const product = JSON.parse(event.body || "{}");
-  const { title, description = "", price = 0, count = 0 } = product;
-
-  if (typeof title !== "string" || title.length === 0) {
-    log.error(`Error while trying to addProduct - title is not defined`);
-    return merge(corsResponse, {
-      statusCode: 400,
-      body: JSON.stringify({
-        ok: false,
-        message: `You should provide the product title`,
-      }),
-    });
-  }
-
-  let productId;
 
   let client;
-  try {
-    client = await PostgresClient.build();
 
+  try {
+    const product = typeof event.body === "string" ? JSON.parse(event.body) : event.body;
+    const { title, description = "", price = 0, count = 0 } = product;
+
+    if (typeof title !== "string" || title.length === 0) {
+      log.error(`Error while trying to addProduct - title is not defined`);
+      return merge(corsResponse, {
+        statusCode: 400,
+        body: JSON.stringify({
+          ok: false,
+          message: `You should provide the product title`,
+        }),
+      });
+    }
+
+    client = await PostgresClient.build();
     await client.beginTransaction();
 
     const responseQueryAddProduct = await client.insertProduct(title, description, price);
@@ -42,7 +41,7 @@ export const addProduct = async event => {
 
     log.info(JSON.stringify(responseQueryAddProduct));
 
-    productId = responseQueryAddProduct.rows[0].id;
+    const productId = responseQueryAddProduct.rows[0].id;
 
     const responseQueryAddStock = await client.insertStocks(productId, count);
     if (!responseQueryAddStock.ok || responseQueryAddStock.rowsAffected === 0) {
@@ -53,26 +52,26 @@ export const addProduct = async event => {
         body: JSON.stringify({ ok: true, message: responseQueryAddStock.message }),
       });
     }
+    await client.endTransaction();
+
+    const imageUrl = `https://source.unsplash.com/featured?nature,water&sig=${productId}`;
+
+    const updatedProduct = { id: productId, ...product, imageUrl };
+
+    return merge(corsResponse, {
+      statusCode: 200,
+      body: JSON.stringify({ ok: true, product: updatedProduct }),
+    });
   } catch (error) {
     log.error(`Error in addProduct call: ${util.inspect(error)}`);
     await client?.rollbackTransaction();
     return merge(corsResponse, {
       statusCode: 500,
-      body: JSON.stringify({ ok: true, message: responseQueryAddStock.message }),
+      body: JSON.stringify({ ok: true, message: error.message }),
     });
   } finally {
-    await client?.endTransaction();
+    await client?.close();
   }
-  await client?.close();
-
-  const imageUrl = `https://source.unsplash.com/featured?nature,water&sig=${productId}`;
-
-  const updatedProduct = { id: productId, ...product, imageUrl };
-
-  return merge(corsResponse, {
-    statusCode: 200,
-    body: JSON.stringify({ ok: true, product: updatedProduct }),
-  });
 
   // Use this code if you don't use the http event with the LAMBDA-PROXY integration
   // return { message: 'Go Serverless v1.0! Your function executed successfully!', event };
