+++ product-service/catalogBatchProcess/catalogBatchProcess.test.js
@@ -1,15 +1,19 @@
-const AWSMock = require("aws-sdk-mock");
-const catalogBatchProcess = require("./handler").catalogBatchProcess;
-const lambdaEventMock = require("lambda-event-mock");
-const log = require("../logger");
+import AWSMock from "aws-sdk-mock";
+import lambdaEventMock from "lambda-event-mock";
+import * as addProduct from "../addProduct";
+import log from "../logger";
+import { catalogBatchProcess } from "./catalogBatchProcess";
 jest.mock("../logger");
-jest.mock("axios");
+jest.mock("../addProduct");
 
 const sqsEventMock = lambdaEventMock
   .sqs()
   .body(JSON.stringify({ id: "TEST_ID", title: "Product1" }))
   .build();
 
+addProduct.addProduct.mockImplementation(() => Promise.resolve({}));
+const logInfoSpy = jest.spyOn(log, "info");
+const logErrorSpy = jest.spyOn(log, "error");
 const snsPublishMock = jest.fn().mockImplementation(() => Promise.resolve());
 
 describe("importFileParser Lambda function test suit:", () => {
@@ -24,14 +28,22 @@ describe("importFileParser Lambda function test suit:", () => {
 
   it("Function not logged any errors", async () => {
     await catalogBatchProcess(sqsEventMock);
-    expect(log.info).toBeCalled();
-    expect(log.error).not.toBeCalled();
+    expect(logInfoSpy).toBeCalled();
+    expect(logErrorSpy).not.toBeCalled();
+  });
+
+  it("Function called addProduct()", async () => {
+    await catalogBatchProcess(sqsEventMock);
+    expect(addProduct.addProduct).toBeCalled();
+    expect(logErrorSpy).not.toBeCalled();
   });
 
   it("Function called sns.Publish", async () => {
     await catalogBatchProcess(sqsEventMock);
     expect(snsPublishMock).toBeCalled();
-    expect(log.info).toBeCalledWith(expect.stringMatching(/title: 'Product1'/));
+    expect(addProduct.addProduct).toBeCalled();
+    expect(logInfoSpy).toBeCalledWith(expect.stringMatching(/title: 'Product1'/));
+    expect(logErrorSpy).not.toBeCalled();
   });
 
   it("Function should log error when sns.Publish throw an error", async () => {
@@ -44,7 +56,7 @@ describe("importFileParser Lambda function test suit:", () => {
     AWSMock.mock("SNS", "publish", snsPublishErrorMock);
     await catalogBatchProcess(sqsEventMock);
     expect(snsPublishErrorMock).toBeCalled();
-    expect(log.error).toBeCalledWith(
+    expect(logErrorSpy).toBeCalledWith(
       expect.stringMatching(/Error while trying to send SNS/)
     );
   });
