+++ product-service/catalogBatchProcess/index.js
@@ -1,0 +1,1 @@
+export { catalogBatchProcess } from "./catalogBatchProcess";
